+++ src/Database/PostgreSQL/PQTypes/Class.hs
@@ -1,7 +1,8 @@
 {-# LANGUAGE OverlappingInstances #-}
 {-# OPTIONS_GHC -fno-warn-deprecated-flags #-}
-module Database.PostgreSQL.PQTypes.Class (
-    MonadDB(..)
+module Database.PostgreSQL.PQTypes.Class
+  ( QueryName(..)
+  , MonadDB(..)
   ) where
 
 import Control.Monad.Trans
@@ -10,6 +11,7 @@ import Control.Monad.Trans.Control
 import Database.PostgreSQL.PQTypes.FromRow
 import Database.PostgreSQL.PQTypes.Internal.Connection
 import Database.PostgreSQL.PQTypes.Internal.Notification
+import Database.PostgreSQL.PQTypes.Internal.Query
 import Database.PostgreSQL.PQTypes.Internal.QueryResult
 import Database.PostgreSQL.PQTypes.SQL.Class
 import Database.PostgreSQL.PQTypes.Transaction.Settings
@@ -20,6 +22,9 @@ class (Applicative m, Monad m) => MonadDB m where
   -- a given time. If simultaneous call is made from another thread, it
   -- will block until currently running 'runQuery' finishes.
   runQuery :: IsSQL sql => sql -> m Int
+  -- | Similar to 'runQuery', but it prepares and executes a statement under a
+  -- given name.
+  runPreparedQuery :: IsSQL sql => QueryName -> sql -> m Int
   -- | Get last SQL query that was executed.
   getLastQuery :: m SomeSQL
   -- | Subsequent queries in the callback do not alter the result of
@@ -75,6 +80,7 @@ instance (
   , MonadDB m
   ) => MonadDB (t m) where
     runQuery = lift . runQuery
+    runPreparedQuery name = lift . runPreparedQuery name
     getLastQuery = lift getLastQuery
     withFrozenLastQuery m = controlT $ \run -> withFrozenLastQuery (run m)
     getConnectionStats = lift getConnectionStats
@@ -85,6 +91,7 @@ instance (
     getNotification = lift . getNotification
     withNewConnection m = controlT $ \run -> withNewConnection (run m)
     {-# INLINE runQuery #-}
+    {-# INLINE runPreparedQuery #-}
     {-# INLINE getLastQuery #-}
     {-# INLINE withFrozenLastQuery #-}
     {-# INLINE getConnectionStats #-}